阶段性重构完成 (1)

+++ kernel/dos/tools/rand.c
@@ -1,11 +1,11 @@
 #include <dos.h>
 #define UINT32_MAX 0xffffffff
-#define INT32_MAX 0x7fffffff
-static unsigned long next1 = 1;
-unsigned int rand(void) {
+#define INT32_MAX  0x7fffffff
+static u32 next1 = 1;
+u32        rand(void) {
   next1 = next1 * 1103515245 + 12345;
-  return ((unsigned int)(next1));
+  return ((u32)(next1));
 }
-void srand(unsigned long seed) {
+void srand(u32 seed) {
   next1 = seed;
 }
